- city manager

+++ BusinessLayer/Managers/CountryManager.cs
@@ -33,17 +33,27 @@ namespace BusinessLayer.Managers
         {
             try
             {
-                var temp = GetAll().FirstOrDefault(s => s.Name == name && s.Continent == continent);
-
-                if (temp == null)
+                var temp = GetAll().FindAll(s => s.Name == name && s.Continent == continent);
+                temp.Remove(con);
+                if (temp.Count == 0)
                 {
-                    con.SetContinent(uow.continentRepo.getById(continent.ID));
-                    con.SetName(name);
-                    con.SetPopulation(pop);
-                    con.SetSurface(sur);
+                    int population = 0;
+                    foreach (var city in con.Cities)
+                        population += city.Population;
+
+                    if (population <= pop)
+                    {
+                        con.SetContinent(uow.continentRepo.getById(continent.ID));
+                        con.SetName(name);
+                        con.SetPopulation(pop);
+                        con.SetSurface(sur);
+                        uow.Complete();
+                    }
+                    else
+                        throw new Exception("Population set to small ");
                 }
                 else
-                    throw new Exception("Country already Exist in Continent" + temp.ID);
+                    throw new Exception("Country already Exist in Continent");
                 RefactorPopulation();
             }
             catch (Exception e)
@@ -55,7 +65,8 @@ namespace BusinessLayer.Managers
         {
             try
             {
-                return uow.countryRepo.getById(id);
+                var temp = uow.countryRepo.getById(id);
+                return temp;
             }
             catch (Exception e)
             {
@@ -77,8 +88,13 @@ namespace BusinessLayer.Managers
         {
             try
             {
-                uow.countryRepo.delete(id);
-                RefactorPopulation();
+                if(Get(id).Cities.Count == 0)
+                {
+                    uow.countryRepo.delete(id);
+                    RefactorPopulation();
+                }
+                else
+                    throw new Exception("Country still contains cities");
             }
             catch (Exception e)
             {
@@ -89,6 +105,10 @@ namespace BusinessLayer.Managers
         {
             try
             {
+                foreach(var country in GetAll())
+                    if (country.Cities.Count != 0)
+                        throw new Exception("there is still a country with cities");
+
                 uow.countryRepo.removeAll();
                 RefactorPopulation();
             }
